Refactor to 3 files
