+++ pkg/controllers/namespace/namespace_controller.go
@@ -17,11 +17,9 @@ limitations under the License.
 package namespace
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
-	"strings"
 	"time"
 
 	openvizapi "go.openviz.dev/apimachinery/apis/openviz/v1alpha1"
@@ -29,33 +27,24 @@ import (
 	"go.openviz.dev/grafana-tools/pkg/controllers/prometheus"
 	"go.openviz.dev/grafana-tools/pkg/detector"
 
-	"github.com/grafana-tools/sdk"
-	v1 "github.com/perses/perses/pkg/model/api/v1"
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	core "k8s.io/api/core/v1"
 	rbac "k8s.io/api/rbac/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	utilerrors "k8s.io/apimachinery/pkg/util/errors"
-	"k8s.io/apimachinery/pkg/util/json"
 	"k8s.io/klog/v2"
-	"k8s.io/utils/ptr"
-	kutil "kmodules.xyz/client-go"
 	kmapi "kmodules.xyz/client-go/api/v1"
 	cu "kmodules.xyz/client-go/client"
 	clustermeta "kmodules.xyz/client-go/cluster"
 	core_util "kmodules.xyz/client-go/core/v1"
-	"kmodules.xyz/client-go/meta"
-	appcatalog "kmodules.xyz/custom-resources/apis/appcatalog/v1alpha1"
 	mona "kmodules.xyz/monitoring-agent-api/api/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
@@ -94,8 +83,6 @@ const (
 )
 
 func (r *ClientOrgReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	log := log.FromContext(ctx)
-
 	var ns core.Namespace
 	if err := r.kc.Get(ctx, req.NamespacedName, &ns); err != nil {
 		return ctrl.Result{}, client.IgnoreNotFound(err)
@@ -351,415 +338,21 @@ func (r *ClientOrgReconciler) Reconcile(ctx context.Context, req ctrl.Request) (
 		}
 	}
 
-	var dashboardList openvizapi.GrafanaDashboardList
-	if err := r.kc.List(ctx, &dashboardList); err != nil {
-		return ctrl.Result{}, err
-	}
-
 	var errList []error
-	for _, dashboard := range dashboardList.Items {
-		if _, isCopy := dashboard.Annotations[srcRefKey]; isCopy || strings.HasSuffix(dashboard.Namespace, "-monitoring") {
-			continue
-		}
-		if dashboard.Spec.Model == nil {
-			return ctrl.Result{}, apierrors.NewBadRequest(fmt.Sprintf("GrafanaDashboard %s/%s is missing a model", dashboard.Namespace, dashboard.Name))
-		}
-		var board sdk.Board
-		err = json.Unmarshal(dashboard.Spec.Model.Raw, &board)
-		if err != nil {
-			return ctrl.Result{}, fmt.Errorf("failed to unmarshal model for GrafanaDashboard %s/%s, reason: %v", dashboard.Namespace, dashboard.Name, err)
-		}
-		for idx, item := range board.Templating.List {
-			if item.Name != "namespace" {
-				continue
-			}
 
-			item.Type = "constant"
-			item.Query = ns.Name
-			board.Templating.List[idx] = item
-		}
-		board.Title = clustermeta.ClientDashboardTitle(board.Title)
-		boardBytes, err := json.Marshal(board)
-		if err != nil {
-			return ctrl.Result{}, err
-		}
-
-		copiedDashboard := &openvizapi.GrafanaDashboard{}
-		err = r.kc.Get(context.TODO(), types.NamespacedName{
-			Namespace: monNamespace.Name,
-			Name:      dashboard.Name,
-		}, copiedDashboard)
-		if err != nil {
-			if !apierrors.IsNotFound(err) {
-				return ctrl.Result{}, err
-			}
-			copiedDashboard = dashboard.DeepCopy()
-			copiedDashboard.Namespace = monNamespace.Name
-		} else {
-			copiedDashboard.Spec = *dashboard.Spec.DeepCopy()
-		}
-
-		opresult, err := cu.CreateOrPatch(ctx, r.kc, copiedDashboard, func(obj client.Object, createOp bool) client.Object {
-			if createOp {
-				copiedDashboard.ResourceVersion = ""
-			}
-
-			if copiedDashboard.Annotations == nil {
-				copiedDashboard.Annotations = map[string]string{}
-			}
-			copiedDashboard.Annotations[srcRefKey] = client.ObjectKeyFromObject(&dashboard).String()
-			copiedDashboard.Annotations[srcHashKey] = meta.ObjectHash(&dashboard)
-
-			// use client Grafana appbinding
-			copiedDashboard.Spec.GrafanaRef = &kmapi.ObjectReference{
-				Namespace: monNamespace.Name,
-				Name:      abClientOrgGrafana,
-			}
-			copiedDashboard.Spec.Model = &runtime.RawExtension{
-				Raw: boardBytes,
-			}
-
-			return copiedDashboard
-		})
-		if err != nil {
-			errList = append(errList, err)
-		} else if opresult != kutil.VerbUnchanged {
-			log.Info(fmt.Sprintf("%s GrafanaDashboard %s/%s", opresult, copiedDashboard.Namespace, copiedDashboard.Name))
-		}
-	}
-
-	var persesDashboardList openvizapi.PersesDashboardList
-	if err := r.kc.List(ctx, &persesDashboardList); err != nil {
+	errListGrafana, err := r.copyGrafanaDashboards(ctx, ns, monNamespace)
+	if err != nil {
 		return ctrl.Result{}, err
 	}
+	errList = append(errList, errListGrafana...)
 
-	for _, dashboard := range persesDashboardList.Items {
-		if dashboard.Spec.Model == nil {
-			return ctrl.Result{}, apierrors.NewBadRequest(fmt.Sprintf("PersesDashboard %s/%s is missing a model", dashboard.Namespace, dashboard.Name))
-		}
-		var board v1.Dashboard
-		err = json.Unmarshal(dashboard.Spec.Model.Raw, &board)
-		if err != nil {
-			return ctrl.Result{}, fmt.Errorf("failed to unmarshal model for GrafanaDashboard %s/%s, reason: %v", dashboard.Namespace, dashboard.Name, err)
-		}
-
-		board.Spec.Display.Name = clustermeta.ClientDashboardTitle(board.Spec.Display.Name)
-		boardBytes, err := json.Marshal(board)
-		if err != nil {
-			return ctrl.Result{}, err
-		}
-
-		if _, found := dashboard.Annotations[srcRefKey]; found && bytes.Equal(dashboard.Spec.Model.Raw, boardBytes) {
-			continue
-		}
-
-		copiedDashboard := &openvizapi.PersesDashboard{}
-		err = r.kc.Get(context.TODO(), types.NamespacedName{
-			Namespace: monNamespace.Name,
-			Name:      dashboard.Name,
-		}, copiedDashboard)
-		if err != nil {
-			if !apierrors.IsNotFound(err) {
-				return ctrl.Result{}, err
-			}
-			copiedDashboard = dashboard.DeepCopy()
-			copiedDashboard.Namespace = monNamespace.Name
-		} else {
-			copiedDashboard.Spec = *dashboard.Spec.DeepCopy()
-		}
-
-		opresult, err := cu.CreateOrPatch(ctx, r.kc, copiedDashboard, func(obj client.Object, createOp bool) client.Object {
-			if createOp {
-				copiedDashboard.ResourceVersion = ""
-			}
-
-			if copiedDashboard.Annotations == nil {
-				copiedDashboard.Annotations = map[string]string{}
-			}
-			copiedDashboard.Annotations[srcRefKey] = client.ObjectKeyFromObject(&dashboard).String()
-			copiedDashboard.Annotations[srcHashKey] = meta.ObjectHash(&dashboard)
-
-			// use client Grafana appbinding
-			copiedDashboard.Spec.PersesRef = &kmapi.ObjectReference{
-				Namespace: monNamespace.Name,
-				Name:      abClientOrgPerses,
-			}
-			copiedDashboard.Spec.Model = &runtime.RawExtension{
-				Raw: boardBytes,
-			}
-
-			return copiedDashboard
-		})
-		if err != nil {
-			errList = append(errList, err)
-		} else if opresult != kutil.VerbUnchanged {
-			log.Info(fmt.Sprintf("%s PersesDashboard %s/%s", opresult, copiedDashboard.Namespace, copiedDashboard.Name))
-		}
+	errListPerses, err := r.copyPersesDashboards(ctx, monNamespace)
+	if err != nil {
+		return ctrl.Result{}, err
 	}
+	errList = append(errList, errListPerses...)
 
 	return ctrl.Result{}, utilerrors.NewAggregate(errList)
-}
-
-// appBindingExists reports whether the named AppBinding is present in the given namespace.
-func (r *ClientOrgReconciler) appBindingExists(ctx context.Context, namespace, name string) bool {
-	var ab appcatalog.AppBinding
-	return r.kc.Get(ctx, types.NamespacedName{Namespace: namespace, Name: name}, &ab) == nil
-}
-
-// setNamespaceMarker stamps the shared registration marker on the namespace.
-func (r *ClientOrgReconciler) setNamespaceMarker(ctx context.Context, monNamespace *core.Namespace, state string) error {
-	rbvt, err := cu.CreateOrPatch(ctx, r.kc, monNamespace, func(in client.Object, _ bool) client.Object {
-		obj := in.(*core.Namespace)
-		if obj.Annotations == nil {
-			obj.Annotations = map[string]string{}
-		}
-		obj.Annotations[prometheus.RegisteredKey] = state
-		return obj
-	})
-	if err != nil {
-		return err
-	}
-	klog.Infof("%s namespace %s with %s annotation", rbvt, monNamespace.Name, prometheus.RegisteredKey)
-	return nil
-}
-
-// registerGrafanaBackend registers the client-org against the Grafana backend and creates its
-// AppBinding. It does not stamp the marker; the caller stamps once both backends succeed.
-func (r *ClientOrgReconciler) registerGrafanaBackend(monNamespace string, pcfg mona.PrometheusConfig, clientOrgId string) error {
-	resp, err := r.bc.Register(mona.PrometheusContext{
-		HubUID:      r.hubUID,
-		ClusterUID:  r.clusterUID,
-		ProjectId:   "",
-		Default:     false,
-		IssueToken:  true,
-		ClientOrgID: clientOrgId,
-	}, pcfg)
-	if err != nil {
-		return err
-	}
-	return r.CreateGrafanaAppBinding(monNamespace, resp)
-}
-
-// registerPersesBackend registers the client-org against the Perses backend and creates its
-// AppBinding. It does not stamp the marker; the caller stamps once both backends succeed.
-func (r *ClientOrgReconciler) registerPersesBackend(monNamespace string, pcfg mona.PrometheusConfig, clientOrgId string) error {
-	persesResp, err := r.bc.RegisterPerses(mona.PrometheusContext{
-		HubUID:      r.hubUID,
-		ClusterUID:  r.clusterUID,
-		ProjectId:   "",
-		Default:     false,
-		IssueToken:  true,
-		ClientOrgID: clientOrgId,
-	}, pcfg)
-	if err != nil {
-		return err
-	}
-	return r.CreatePersesAppBinding(monNamespace, persesResp)
-}
-
-func (r *ClientOrgReconciler) CreateGrafanaAppBinding(monNamespace string, resp *prometheus.GrafanaDatasourceResponse) error {
-	ab := appcatalog.AppBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      abClientOrgGrafana,
-			Namespace: monNamespace,
-		},
-	}
-
-	abvt, err := cu.CreateOrPatch(context.TODO(), r.kc, &ab, func(in client.Object, createOp bool) client.Object {
-		obj := in.(*appcatalog.AppBinding)
-
-		//ref := metav1.NewControllerRef(prom, schema.GroupVersionKind{
-		//	Group:   monitoring.GroupName,
-		//	Version: monitoringv1.Version,
-		//	Kind:    "Prometheus",
-		//})
-		//obj.OwnerReferences = []metav1.OwnerReference{*ref}
-		//
-		//if obj.Annotations == nil {
-		//	obj.Annotations = make(map[string]string)
-		//}
-		//obj.Annotations["monitoring.appscode.com/is-default-grafana"] = "true"
-
-		obj.Spec.Type = "Grafana"
-		obj.Spec.AppRef = nil
-		obj.Spec.ClientConfig = appcatalog.ClientConfig{
-			URL: ptr.To(resp.Grafana.URL),
-			//Service: &appcatalog.ServiceReference{
-			//	Scheme:    "http",
-			//	Namespace: svc.Namespace,
-			//	Name:      svc.Name,
-			//	Port:      0,
-			//	Path:      "",
-			//	Query:     "",
-			//},
-			//InsecureSkipTLSVerify: false,
-			//CABundle:              nil,
-			//ServerName:            "",
-		}
-		obj.Spec.Secret = &appcatalog.TypedLocalObjectReference{
-			APIGroup: "",
-			Kind:     "Secret",
-			Name:     ab.Name + "-auth",
-		}
-
-		// TODO: handle TLS config returned in resp
-		if caCert := r.bc.CACert(); len(caCert) > 0 {
-			obj.Spec.ClientConfig.CABundle = caCert
-		}
-
-		params := openvizapi.GrafanaConfiguration{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "GrafanaConfiguration",
-				APIVersion: openvizapi.SchemeGroupVersion.String(),
-			},
-			Datasource: resp.Datasource,
-			FolderID:   resp.FolderID,
-		}
-		paramBytes, err := json.Marshal(params)
-		if err != nil {
-			panic(err)
-		}
-		obj.Spec.Parameters = &runtime.RawExtension{
-			Raw: paramBytes,
-		}
-
-		return obj
-	})
-	if err == nil {
-		klog.Infof("%s AppBinding %s/%s", abvt, ab.Namespace, ab.Name)
-
-		authSecret := core.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      ab.Name + "-auth",
-				Namespace: monNamespace,
-			},
-		}
-
-		svt, e2 := cu.CreateOrPatch(context.TODO(), r.kc, &authSecret, func(in client.Object, createOp bool) client.Object {
-			obj := in.(*core.Secret)
-
-			ref := metav1.NewControllerRef(&ab, schema.GroupVersionKind{
-				Group:   appcatalog.SchemeGroupVersion.Group,
-				Version: appcatalog.SchemeGroupVersion.Version,
-				Kind:    "AppBinding",
-			})
-			obj.OwnerReferences = []metav1.OwnerReference{*ref}
-
-			obj.StringData = map[string]string{
-				"token": resp.Grafana.BearerToken,
-			}
-
-			return obj
-		})
-		if e2 == nil {
-			klog.Infof("%s Grafana auth secret %s/%s", svt, authSecret.Namespace, authSecret.Name)
-		}
-	}
-
-	return err
-}
-
-func (r *ClientOrgReconciler) CreatePersesAppBinding(monNamespace string, resp *prometheus.PersesDatasourceResponse) error {
-	ab := appcatalog.AppBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      abClientOrgPerses,
-			Namespace: monNamespace,
-		},
-	}
-
-	abvt, err := cu.CreateOrPatch(context.TODO(), r.kc, &ab, func(in client.Object, createOp bool) client.Object {
-		obj := in.(*appcatalog.AppBinding)
-
-		//ref := metav1.NewControllerRef(prom, schema.GroupVersionKind{
-		//	Group:   monitoring.GroupName,
-		//	Version: monitoringv1.Version,
-		//	Kind:    "Prometheus",
-		//})
-		//obj.OwnerReferences = []metav1.OwnerReference{*ref}
-		//
-		//if obj.Annotations == nil {
-		//	obj.Annotations = make(map[string]string)
-		//}
-		//obj.Annotations["monitoring.appscode.com/is-default-grafana"] = "true"
-
-		obj.Spec.Type = "Perses"
-		obj.Spec.AppRef = nil
-		obj.Spec.ClientConfig = appcatalog.ClientConfig{
-			URL: ptr.To(resp.Perses.URL),
-			//Service: &appcatalog.ServiceReference{
-			//	Scheme:    "http",
-			//	Namespace: svc.Namespace,
-			//	Name:      svc.Name,
-			//	Port:      0,
-			//	Path:      "",
-			//	Query:     "",
-			//},
-			//InsecureSkipTLSVerify: false,
-			//CABundle:              nil,
-			//ServerName:            "",
-		}
-		obj.Spec.Secret = &appcatalog.TypedLocalObjectReference{
-			Name:     ab.Name + "-auth",
-			APIGroup: "",
-			Kind:     "Secret",
-		}
-
-		// TODO: handle TLS config returned in resp
-		if caCert := r.bc.CACert(); len(caCert) > 0 {
-			obj.Spec.ClientConfig.CABundle = caCert
-		}
-
-		params := openvizapi.PersesConfiguration{
-			TypeMeta: metav1.TypeMeta{
-				Kind:       "PersesConfiguration",
-				APIVersion: openvizapi.SchemeGroupVersion.String(),
-			},
-			Datasource:  resp.Datasource,
-			FolderName:  resp.FolderName,
-			ProjectName: resp.ProjectName,
-		}
-		paramBytes, err := json.Marshal(params)
-		if err != nil {
-			panic(err)
-		}
-		obj.Spec.Parameters = &runtime.RawExtension{
-			Raw: paramBytes,
-		}
-
-		return obj
-	})
-	if err == nil {
-		klog.Infof("%s AppBinding %s/%s", abvt, ab.Namespace, ab.Name)
-
-		authSecret := core.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      ab.Name + "-auth",
-				Namespace: monNamespace,
-			},
-		}
-
-		svt, e2 := cu.CreateOrPatch(context.TODO(), r.kc, &authSecret, func(in client.Object, createOp bool) client.Object {
-			obj := in.(*core.Secret)
-
-			ref := metav1.NewControllerRef(&ab, schema.GroupVersionKind{
-				Group:   appcatalog.SchemeGroupVersion.Group,
-				Version: appcatalog.SchemeGroupVersion.Version,
-				Kind:    "AppBinding",
-			})
-			obj.OwnerReferences = []metav1.OwnerReference{*ref}
-
-			obj.StringData = map[string]string{
-				"token": resp.Perses.BearerToken,
-			}
-
-			return obj
-		})
-		if e2 == nil {
-			klog.Infof("%s Grafana auth secret %s/%s", svt, authSecret.Namespace, authSecret.Name)
-		}
-	}
-
-	return err
 }
 
 // SetupWithManager sets up the controller with the Manager.
